Combine SignalTestLaws with SignalSpec
This brings the Signal laws in line with the rest of the test code, e.g.
ChunkSpec.
